[backend] fix hanging bug after pytest
for some python versions (3.13.x+) we were not terminating correctly,
causing eg pytest to hang forever

we fix that by correctly sending SIGINT first at the graceful exit,
because some python multiprocessing resources require that for
a clean exit (as opposed to just SIGTERM)

we additionally register a SIGTERM handler in the launcher app,
to react to an outer SIGTERM correctly as well

lastly, we reduce the log pollution during pytests due to findlibs

+++ backend/forecastbox/standalone/procs.py
@@ -4,6 +4,8 @@
 """
 
 import logging
+import os
+import signal
 from dataclasses import dataclass
 from multiprocessing import Process, connection
 
@@ -22,9 +24,16 @@ class ChildProcessGroup:
 
     def shutdown(self):
         for p in self.procs:
-            p.terminate()
-            p.join(1)
-            p.kill()
+            if p.is_alive():
+                # p.interrupt() # TODO after 3.14 add
+                os.kill(p.pid, signal.SIGINT)
+                p.join(3)
+            if p.is_alive():
+                p.terminate()
+                p.join(3)
+            if p.is_alive():
+                p.kill()
+                p.join(3)
 
 
 def previous_cleanup():
